Separated threads from restnode.py + beginning of PoS
